pyproject.toml
[project]
name = "interpretable_bioacoustic_classifiers"
version = "1.0.0"
description = "self-supervised learning of representations for ecoacoustic inference"
authors = [
  {name = "Kieran Gibb", email =  "hello@kierangibb.co.uk"},
  {name = "Ivor Simpson", email =  "i.simpson@sussex.ac.uk"},
  {name = "Alice Eldridge", email =  "alicee@sussex.ac.uk"},
]
requires-python = ">=3.10,<3.12"

dependencies = [
  "librosa>=0.10.1",
  "lightning==2.2.1",
  "matplotlib>=3.4.3",
  "pyarrow>=15.0.2",
  "torch==2.4.1",
  "torchaudio>=2.4.1",
  "torchvision>=0.19.1",
  "seaborn>=0.13.2",
  "jupyter-book>=0.15.1",
  "jupyterlab>=4.3.5",
  "jupyter-contrib-nbextensions>=0.7.0",
  "python-dotenv>=1.0.1",
  "ipywidgets>=8.1.5",
  "ipympl>=0.9.6",
  "anywidget>=0.9.13",
  "jupyterlab-widgets>=3.0.13",
  "birdnet==0.1.6",
  "wandb>=0.21.0",
  "torchmetrics>=0.11.4",
  "hydra-core==1.3.2",
  "hydra-colorlog==1.2.0",
  "hydra-optuna-sweeper==1.2.0",
  "rootutils==1.0.7",
  "tabulate>=0.9.0",
  "dash>=3.2.0",
  "dash-mantine-components>=2.4.0",
  "cmasher>=1.9.2",
  "hydra-joblib-launcher>=1.2.0",
  "ranzen[all]>=2.5.1,<3.0.0",
  "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
  "pytest==7.1.3",
  "black==22.10.0",
  "isort==5.10.1",
]

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  {index = "pytorch-gpu"}
]
torchvision = [
    {index = "pytorch-gpu"}
]
torchaudio = [
    {index = "pytorch-gpu"}
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "test/"
